Module Tezos_dal_node_lib

module Dal_node_client : sig ... end

module Dal_plugin : sig ... end

module Dal_proto_types : sig ... end

module Dal_store_sqlite3 : sig ... end

module Dal_store_sqlite3_events : sig ... end

module Kvs_skip_list_cells_store : sig ... end

This module instantiates the key value store to provide facilities for storing and retrieving the cells of the DAL skip list. The store maintains:

module Store_migrations : sig ... end

module Value_size_hooks : sig ... end

The encoded_value_size required by Key_value_store.layout is known when the daemon loads a protocol, after the store is initialized. We use the closures like set_share_size to perform delayed protocol-specific parameter passing.