Module Tezos_dal_node_lib.Dal_node_client

type cctxt

Instance of Tezos_client_base.Client_context that only handles IOs and RPCs. Can be used for keys and RPCs related commands.

val call : cctxt -> ([< Resto.meth ], unit, 'a, 'b, 'c, 'd, Tezos_error_monad.Error_monad.error list) Resto.MakeService(Tezos_rpc__.RPC_encoding).t -> 'a -> 'b -> 'c -> 'd Tezos_error_monad.Error_monad.tzresult Lwt.t

val make_unix_cctxt : Uri.t -> cctxt

make_unix_context endpoint generates a cctxt with the provided endpoint.

val get_slot_pages : cctxt -> Tezos_dal_node_services.Types.slot_id -> bytes list Tezos_base.TzPervasives.tzresult Lwt.t

get_slot_pages cctxt slot_id fetches from the dal node the pages of the slot with id slot_id.

val get_slot_page_proof : cctxt -> Tezos_dal_node_services.Types.slot_id -> int -> Tezos_crypto_dal.Cryptobox.page_proof Tezos_base.TzPervasives.tzresult Lwt.t

get_slot_page_proof cctxt slot_id page_index computes and returns the proof of the page whose index and slot are given.

val post_slot : cctxt -> ?slot_index:Tezos_dal_node_services.Types.slot_index -> string -> (Tezos_crypto_dal.Cryptobox.commitment * Tezos_crypto_dal.Cryptobox.commitment_proof) Tezos_base.TzPervasives.tzresult Lwt.t

post_slot cctxt slot_data posts the given data as a slot to the DAL node, and returns the corresponding commitment hash alongside its proof.