Module Tezos_dal_node_lib.Dal_plugin

type operation_application_result =

| Succeeded
(*
The associated operation is included in a block and successfully applied.
*)

| Failed
(*
The associated operation is included in a block but its application was not successful (failed, backtracked or skipped).
*)

type slot_index = int

type slot_header = {

published_level : int32;

slot_index : slot_index;

commitment : Tezos_crypto_dal.Cryptobox.Verifier.commitment;

}

Information extracted from DAL slots headers operations included in L1 blocks. Each slot header is made of an L1 level for which it is published, the slot's index and commitment.

type proto_parameters = {

feature_enable : bool;

incentives_enable : bool;

number_of_slots : int;

attestation_lag : int;

attestation_threshold : int;

cryptobox_parameters : Tezos_crypto_dal.Cryptobox.Verifier.parameters;

sc_rollup_challenge_window_in_blocks : int;

commitment_period_in_blocks : int;

dal_attested_slots_validity_lag : int;

blocks_per_cycle : int32;

}

val proto_parameters_encoding : proto_parameters Tezos_base.TzPervasives.Data_encoding.t

module type T = sig ... end

val register : (module T) -> unit

val get : Tezos_base.TzPervasives.Protocol_hash.Table.key -> (module T) option