Module type Dal_plugin.T

module Proto : Tezos_protocol_updater.Registered_protocol.T

type block_info

type dal_attestation

val block_info : ?chain:Tezos_shell_services.Block_services.chain -> ?block:Tezos_shell_services.Block_services.block -> metadata:[ `Always | `Never ] -> Tezos_rpc.Context.generic -> block_info Tezos_base.TzPervasives.tzresult Lwt.t

block_info ?chain ?block ~metadata ctxt returns the information of the block in ctxt for the given chain. Block's metadata are included or skipped depending on the value of metadata. This is a wrapper on top of Protocol_client_context.Alpha_block_services.info.

val get_constants : Tezos_shell_services.Chain_services.chain -> Tezos_shell_services.Block_services.block -> Tezos_rpc.Context.generic -> proto_parameters Tezos_base.TzPervasives.tzresult Lwt.t

val get_published_slot_headers : block_info -> (slot_header * operation_application_result) list Tezos_base.TzPervasives.tzresult Lwt.t

val get_dal_content_of_attestations : block_info -> (int * Tezos_base.TzPervasives.Signature.Public_key_hash.t option * dal_attestation option) list

For a given block, returns for each included attestation, as a list, its Tenderbake slot, its attester (if available in the operation receipt), and its DAL attestation.

val get_committee : Tezos_rpc.Context.generic -> level:int32 -> int list Tezos_base.TzPervasives.Signature.Public_key_hash.Map.t Tezos_base.TzPervasives.tzresult Lwt.t

get_committee ctxt ~level retrieves the DAL committee at level from L1 as a map that associates to the public key hash pkh of the member of the committee its assigned shard indexes.

val dal_attestation : block_info -> dal_attestation Tezos_base.TzPervasives.tzresult

dal_attestation block_info returns the metadata of the given block_info as an abstract value of type dal_attestation to be passed to the is_attested function.

Fails with Cannot_read_block_metadata if block_info's metadata are stripped.

val is_attested : dal_attestation -> slot_index -> bool

is_attested dal_attestation index returns true if index is one of the dal_attestation and false otherwise.

val get_round : Tezos_base.TzPervasives.Fitness.t -> int32 Tezos_base.TzPervasives.tzresult

get_round fitness returns the block round contained in fitness.

val block_shell_header : block_info -> Tezos_base.TzPervasives.Block_header.shell_header

block_shell_header block_info returns the shell header of the block whose information are given .

module Skip_list : sig ... end

module RPC : sig ... end