Module Dal_proto_types.Skip_list_hash

Generic protocol-agnostic representation of DAL skip list hashes.

type t

The generic protocol agnostic representation of a protocol type.

val of_proto : 'a Tezos_base.TzPervasives.Data_encoding.t -> 'a -> t

Import a value of the protocol value into the generic type.

val to_proto : 'a Tezos_base.TzPervasives.Data_encoding.t -> t -> 'a

Dual operation of of_proto.

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val equal : t -> t -> bool