Module Dal_store_sqlite3.Skip_list_cells

type t

A handler to the DAL node's skip list cells SQLite3 database.

val init : data_dir:string -> perm:[ `Read_only | `Read_write ] -> unit -> t Tezos_base.TzPervasives.tzresult Lwt.t

init ~data_dir ~perm () returns a handler to the DAL node store located under data_dir. If no store is located in data_dir, an empty store is created.

If perm is `Read_only, then SQL requests requiring write access will fail. With `Read_write, they will succeed as expected.

val close : t -> unit Lwt.t

close t closes the store by freeing all resources and closing database connections.

val use : t -> (conn -> 'a Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a Tezos_base.TzPervasives.tzresult Lwt.t

use t k executes k with a fresh connection to t.

val find : ?conn:conn -> t -> Dal_proto_types.Skip_list_hash.t -> Dal_proto_types.Skip_list_cell.t Tezos_base.TzPervasives.tzresult Lwt.t

find ?conn store hash returns the cell associated to hash in the store, if any. Uses the conn if provided (defaults to None).

val insert : ?conn:conn -> t -> attested_level:int32 -> (Dal_proto_types.Skip_list_hash.t * Dal_proto_types.Skip_list_cell.t) list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

insert ?conn store ~attested_level values inserts the given list of values associated to the given attested_level in the store. Any existing value is overridden. Uses the conn if provided (defaults to None).

val remove : ?conn:conn -> t -> attested_level:int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

remove ?conn store ~attested_level removes any data related to attested_level from the store. Uses the conn if provided (defaults to None).

val schemas : t -> string list Tezos_base.TzPervasives.tzresult Lwt.t

schemas t returns the list of SQL statements allowing to recreate the tables of the store t.

module Internal_for_tests : sig ... end

Internal functions for testing purpose.