Module Kvs_skip_list_cells_store.Internal_for_migrations

Internal functions for storage migrations.

val get_attested_levels : t -> int32 Lwt_stream.t

get_attested_levels store returns the attested levels registered in the store. Since the number of attested levels can be large, it returns a stream.

val find_hash : t -> attested_level:int32 -> slot_index:int -> Dal_proto_types.Skip_list_hash.t Tezos_base.TzPervasives.tzresult Lwt.t

find_hash store ~attested_level ~slot_index returns the cell hash associated to the attested_level and the slot_index in the store, if any.