Module Tezos_dal_node_lib_gossipsub.Gossipsub

This module exposes the instantiations of the Gossipsub and Octez-p2p libraries to be used by the DAL node to connect to and exchange data with peers.

module Worker : sig ... end

The worker module exposes instantiation of the Gossipsub worker functor, alongside the config used to instantiate the functor and the default values of the GS parameters.

module Transport_layer : sig ... end

The transport layer module exposes the needed primitives, interface and default parameters for the instantiation of the Octez-p2p library.

module Transport_layer_hooks : sig ... end

This module implements the list of hooks that allow interconnecting the Gossipsub worker with the transport layer. They are exposed via the Transport_layer_hooks.activate function below.

val version : network_name:string -> Tezos_base.TzPervasives.Network_version.t

version ~network_name returns the current version of the P2P.