Module Gossipsub.Transport_layer_hooks

This module implements the list of hooks that allow interconnecting the Gossipsub worker with the transport layer. They are exposed via the Transport_layer_hooks.activate function below.

val activate : Worker.t -> Transport_layer.t -> app_messages_callback: (Tezos_dal_node_services.Types.Message.t -> Tezos_dal_node_services.Types.Message_id.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t) -> unit Lwt.t

See Gs_transport_connection.activate.