Module Worker.Validate_message_hook

A hook to set or update messages and messages IDs validation function. Should be called once at startup and every time the DAL parameters change.

val set : (?message:GS.Message.t -> message_id:GS.Message_id.t -> unit -> [ `Valid | `Unknown | `Outdated | `Invalid ]) -> unit