Module Tezos_dal_node_lib_gossipsub.Gs_transport_connection

val activate : Gs_interface.Worker_instance.t -> (Gs_interface.Worker_instance.p2p_message, Tezos_dal_node_services.Types.P2P.Metadata.Peer.t, Tezos_dal_node_services.Types.P2P.Metadata.Connection.t) Tezos_p2p.P2p.t -> app_messages_callback: (Tezos_dal_node_services.Types.Message.t -> Tezos_dal_node_services.Types.Message_id.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t) -> unit Lwt.t

activate gs_worker transport_layer ~app_messages_handler connects the given gs_worker and transport_layer. (Dis)connections and messages of the transport layer are forwarded to the GS worker. P2P output messages and (dis)connection requests are forwarded from the GS worker to the transport layer.

The app_messages_handler is invoked when some application messages are put by the Gossipsub worker in the application output stream.