Module Tezos_dal_node_lib_gossipsub.Transport_layer_interface

module Types = Tezos_dal_node_services.Types

This module defines the data structures used to instantiate the Octez P2P library. More exactly, it exposes:

the types p2p_message, peer_metadata and connection_metadata, used to instantiate type variables of P2p.t;

the values peer_metadata_cfg of type P2p_params.peer_meta_config, conn_metadata_cfg of type P2p_params.conn_meta_config and message_cfg of type P2p_params.message_config that are used when calling P2p.create.

type px_peer = Types.Peer.t

Peers exchanged via PX. point represents the (address, port) pair of the exchanged peer, while peer represents the cryptographic identity of the peer.

val p2p_message_encoding : Gs_interface.Worker_instance.p2p_message Tezos_base.TzPervasives.Data_encoding.t

val pp_p2p_message : Stdlib.Format.formatter -> Gs_interface.Worker_instance.p2p_message -> unit

val message_config : network_name:string -> Gs_interface.Worker_instance.p2p_message Tezos_base.TzPervasives.P2p_params.message_config

A P2P message config is parameterized by the network's name.

val version : network_name:string -> Tezos_base.TzPervasives.Network_version.t